requirements.txt
aiofiles==0.6.0
aiohttp==3.6.2
astral==1.10.1
astroid==2.4.2
async-timeout==3.0.1
Automat==20.2.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
ciso8601==2.1.3
click==7.1.2
cmake==3.18.2.post1
constantly==15.1.0
cryptography==2.9.2
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
homeassistant==0.116.4
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
isort==5.6.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
mypy==0.790
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.19.2
observable==1.0.3
opencv-python-headless==4.4.0.46
Pillow==8.0.1
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.15.100
sine.properties==0.1.1
singleton-decorator==1.0.0
text-unidecode==1.3
twilio==6.50.1
Twisted==20.3.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.11
voluptuous==0.12.0
voluptuous-serialize==2.4.0
wrapt==1.12.1
yarl==1.4.2
zope.interface==5.2.0
